requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
gunicorn==19.7.1
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
django-tinymce4-lite
